영화 관리에서 상영 내용이 있는 부분 삭제 막기

+++ khFinalproject/src/main/java/com/kh/spring/controller/MovieController.java
@@ -199,8 +199,15 @@ public class MovieController {
 	
 	@GetMapping("/admin/delete")
 	public String delete(@RequestParam int movieNo) {
-		movieService.delete(movieNo);
-		return "redirect:/movie/admin/list";
+		
+		LastInfoViewDto checkDto = lastInfoViewDao.exist(movieNo);
+		if(checkDto == null) {
+			movieService.delete(movieNo);
+			return "redirect:/movie/admin/list";
+		}else {
+			return "redirect:/movie/admin/list?error";
+		}
+		
 	}
 	
 	@GetMapping("/admin/edit")
